+++ packages/wordpress-plugin-boilerplate/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+### Changed
+
+* Don't rename plugin main file. See [wearerequired/wordpress-plugin-boilerplate#133](https://github.com/wearerequired/wordpress-plugin-boilerplate/pull/133).
+
 ## [0.1.2] - 2021-03-17
 
 ### Changed
@@ -23,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Initial release.
 
-[Unreleased]:https://github.com/wearerequired/js/compare/@wearerequired/wordpress-plugin-boilerplate@0.1.3...HEAD
+[Unreleased]:https://github.com/wearerequired/js/compare/@wearerequired/wordpress-plugin-boilerplate@0.1.2...HEAD
 [0.1.2]: https://github.com/wearerequired/js/compare/@wearerequired/wordpress-plugin-boilerplate@0.1.1...@wearerequired/wordpress-plugin-boilerplate@0.1.2
 [0.1.1]: https://github.com/wearerequired/js/compare/@wearerequired/wordpress-plugin-boilerplate@0.1.0...@wearerequired/wordpress-plugin-boilerplate@0.1.1
 [0.1.0]: https://github.com/wearerequired/js/releases/tag/@wearerequired/wordpress-plugin-boilerplate@0.1.0